move funcs to faccess and use AsRawHandle
